+++ SmbAbstraction/Exception/InvalidCredentialException.cs
@@ -2,8 +2,13 @@ using System;
 
 namespace SmbAbstraction
 {
+    [Serializable]
     public class InvalidCredentialException : Exception
     {
+        public InvalidCredentialException() : base()
+        {
+        }
+
         public InvalidCredentialException(string message) : base(message)
         {
         }